Adding Documentation
Two files have been added as documentation. An intro file as another explanation of the project and a start file to explain what I am currently working on and to list version history. Version history may be moved to its own file in the future so it can be found easier.